Struct mouse_sovec::SoVec

pub struct SoVec<T, A> where
    A: GlobalAlloc,  { /* fields omitted */ }

SoVec stands for Small optimized Vector .

SoVec behaves like std::collections::Vec, however, it will not allocate heap memory if the requested memory size is small enough. Instead of heap, it uses itself as a buffer then.

To avoid allocating as much as possible, the performance is better than that of std::collections::Vec .

Some unsafe methods behavior is different from the same name method of std::collections::Vec ; otherwise, (i.e. the all safe methods) behaves like that of std::collections::Vec .

Implementations

impl<T, A> SoVec<T, A> where
A: GlobalAlloc,

pub fn new(alloc: A) -> Self

Creates a new empty instance.

pub fn with_capacity(capacity: usize, alloc: A) -> Self

Creates a new empty instance whose capacity is greater than or equals to capacity .

pub fn len(&self) -> usize

Returns the number of the elements self is holding.

pub fn is_empty(&self) -> bool

Returns true if self is not holding any element, or false.

pub unsafe fn set_len(&mut self, new_len: usize)

Forces the length of self to new\_len .

Safety

new\_len must be less than or equal to capacity .

The elements at old_len..new_len must be initialized when extending.

The elements at new_len..old_len must be dropped when shrinking.

pub fn capacity(&self) -> usize

Returns the number of the elements self can hold without allocating.

pub fn reserve_exact(&mut self, additional: usize)

Reserves the minimum capacity to insert additional more elements.

After this method is called, self.capacity will return the number to be greater than or equals to self.len() + additional .

pub unsafe fn push(&mut self, elm: T)

Appends elm to the end of self .

The caller must ensure that self has sufficient capacity in advance.

Safety

The behavior is undefined if self.len is greater than or equals to self.capacity .

pub fn pop(&mut self) -> Option<T>

Removes the last element and returns it if any.

pub unsafe fn as_ptr(&self) -> *const T

Returns a raw pointer to the buffer of self .

Warnings

The caller must ensure that self outlives the pointer this function returns, or else it will end up pointing to garbage. Modifying or moving self may cause its buffer to be reallocated, which would also make any pointers to it invalid.

Safety

This method itself is safe, however, moving self can invalidate the returned value unlike to the same name methods of std::collections::Vec .

unsafe modifier is added to alert it.

pub unsafe fn as_mut_ptr(&mut self) -> *mut T

Returns a raw pointer to the buffer of self .

Warnings

The caller must ensure that self outlives the pointer this function returns, or else it will end up pointing to garbage. Modifying or moving self may cause its buffer to be reallocated, which would also make any pointers to it invalid.

Safety

This method itself is safe, however, moving self can invalidate the returned value unlike to the same name methods of std::collections::Vec .

unsafe modifier is added to alert it.

pub fn clear(&mut self)

Removes the all elements keeping the allocated capacity, and set the length 0.

Note this has the same effect to self.truncate(0) .

pub fn truncate(&mut self, new_len: usize)

Enshortens self, keeping the first new_len elements and dropping the rest.

If new_len is greater than or equals to the current length, nothing is done.

Note that this method does not have any effect on the allocated capacity of self .

pub fn shrink_to_fit(&mut self)

Shrinks the capacitance of self as much as possible.